Change CPDF_NameTree::LookupValueAndName() to take a size_t.
Pass a size_t index through to the internal implementation. Do the same
for DeleteValueAndName(). All callers have been audited. They either
pass in size_t already, or checks int indices to make sure they are not
negative.

+++ core/fpdfdoc/cpdf_nametree.cpp
@@ -482,7 +482,7 @@ bool CPDF_NameTree::AddValueAndName(RetainPtr<CPDF_Object> pObj,
   return true;
 }
 
-bool CPDF_NameTree::DeleteValueAndName(int nIndex) {
+bool CPDF_NameTree::DeleteValueAndName(size_t nIndex) {
   absl::optional<IndexSearchResult> result =
       SearchNameNodeByIndex(m_pRoot.Get(), nIndex);
   if (!result) {
@@ -500,7 +500,7 @@ bool CPDF_NameTree::DeleteValueAndName(int nIndex) {
   return true;
 }
 
-CPDF_Object* CPDF_NameTree::LookupValueAndName(int nIndex,
+CPDF_Object* CPDF_NameTree::LookupValueAndName(size_t nIndex,
                                                WideString* csName) const {
   absl::optional<IndexSearchResult> result =
       SearchNameNodeByIndex(m_pRoot.Get(), nIndex);
